every friend enpoint is working now

+++ app/friends/routes.py
@@ -1,7 +1,8 @@
 from flask import request, jsonify, Blueprint, current_app
 from sqlalchemy import select
-import uuid
+from uuid import uuid4 as uuid
 
+# import uuid
 from app.extensions import db
 from .models import Friend, FriendRequest
 
@@ -9,12 +10,11 @@ from .models import Friend, FriendRequest
 def add_routes(bp: Blueprint):
     @bp.post("/friend_request")
     def friend_request():
-        # add POST route to save a new entry in the friend request table, taking in the user id of the person who sent the request and the user id of the person who received the request
+        # add POST route to save a new entry in the friend request table, taking in the user id of the
+        # person who sent the request and the user id of the person who received the request
+        # I get an error, module' object is not callable, can you change the following code to fix this
         post_data = request.get_json()
         try:
-            from_user_id = post_data.get("from_user_id")
-            to_user_id = post_data.get("to_user_id")
-
             friend_request = FriendRequest(
                 id=uuid(),
                 from_user_id=post_data.get("from_user_id"),
@@ -40,12 +40,20 @@ def add_routes(bp: Blueprint):
         try:
             from_user_id = post_data.get("from_user_id")
             to_user_id = post_data.get("to_user_id")
-            friend_request = FriendRequest.query.filter_by(
-                from_user_id=from_user_id, to_user_id=to_user_id
+            friend_request = db.session.scalars(
+                select(FriendRequest).filter_by(
+                    from_user_id=from_user_id, to_user_id=to_user_id
+                )
             ).first()
+            if friend_request is None:
+                response_object = {
+                    "status": "fail",
+                    "message": "Friend request does not exist",
+                }
+                return jsonify(response_object), 401
             db.session.delete(friend_request)
             db.session.commit()
-            friend = Friend(id=uuid(), from_user_id=from_user_id, to_user_id=to_user_id)
+            friend = Friend(id=uuid(), user_a_id=from_user_id, user_b_id=to_user_id)
             db.session.add(friend)
             db.session.commit()
             response_object = {
@@ -65,9 +73,17 @@ def add_routes(bp: Blueprint):
         try:
             from_user_id = post_data.get("from_user_id")
             to_user_id = post_data.get("to_user_id")
-            friend_request = FriendRequest.query.filter_by(
-                from_user_id=from_user_id, to_user_id=to_user_id
+            friend_request = db.session.scalars(
+                select(FriendRequest).filter_by(
+                    from_user_id=from_user_id, to_user_id=to_user_id
+                )
             ).first()
+            if friend_request is None:
+                response_object = {
+                    "status": "fail",
+                    "message": "Friend request does not exist",
+                }
+                return jsonify(response_object), 401
             db.session.delete(friend_request)
             db.session.commit()
             response_object = {
@@ -80,15 +96,28 @@ def add_routes(bp: Blueprint):
             response_object = {"status": "fail", "message": str(exception)}
             return jsonify(response_object), 503
 
-    @bp.get("/friend/delete")
+    @bp.post("/friend/delete")
     def delete_friend():
         # GET request to remove a friend from the friends table
+        post_data = request.get_json()
         try:
-            from_user_id = request.args.get("user_a")
-            to_user_id = request.args.get("user_b")
-            friend = Friend.query.filter_by(
-                from_user_id=from_user_id, to_user_id=to_user_id
+            from_user_id = post_data.get("user_a_id")
+            to_user_id = post_data.get("user_b_id")
+            friend = db.session.scalars(
+                select(Friend).filter_by(user_a_id=from_user_id, user_b_id=to_user_id)
             ).first()
+            if friend is None:
+                friend = db.session.scalars(
+                    select(Friend).filter_by(
+                        user_a_id=to_user_id, user_b_id=from_user_id
+                    )
+                ).first()
+                if friend is None:
+                    response_object = {
+                        "status": "fail",
+                        "message": "Friend does not exist",
+                    }
+                    return jsonify(response_object), 401
             db.session.delete(friend)
             db.session.commit()
             response_object = {
